requirements.txt
absl-py==2.2.0
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
humanfriendly==10.0
idna==3.10
joblib==1.4.2
keras==3.9.0
keyboard==0.13.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
numpy==2.1.3
onnxruntime==1.21.0
openwakeword==0.6.0
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
protobuf==5.29.4
PyAudio==0.2.14
Pygments==2.19.1
pyreadline3==3.5.4
pyserial==3.5
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.0.1
simpleaudio==1.0.4
six==1.17.0
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
tflite==2.18.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
